requirements.txt
accelerate==1.11.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bert-score==0.3.13
bleach==6.3.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
docopt==0.6.2
durationpy==0.10
effdet==0.4.1
emoji==2.15.0
executing==2.2.1
faiss-cpu==1.12.0
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-vision==3.11.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.0.3
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-experimental==0.0.42
langchain-huggingface==1.0.1
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
nltk==3.9.2
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.2
openai==2.7.1
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.1
pickleshare==0.7.5
pikepdf==10.0.0
pillow==12.0.0
pipreqs==0.5.0
platformdirs==4.5.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.3.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.0
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0